Revert "Don't rely so much on tar(1)."
This reverts commit a6d7852d2a14de438c5a386a083228d08fef621b.

Because of limitations in certain containerized situations, I changed
from using `tar` to using `cp --archive` to copy file hierarchies in two
commits.  One was ed15dd16, and that has already been reverted because
it broke on Mac OS X.  Here, revert the other one; it would break on Mac
just like the other one, and we don't need it because we aren't building
in containerized environments any more and don't expect to have to do so
going forward.