Add support for a generic POD type
This lets you write things like:

```chpl
proc myproc(arg: POD) {
  writeln("A POD type");
}
```

The advantage of doing this over something like:

```chpl
proc myproc(arg) where isPODType(arg) {
  writeln("A POD type");
}
```

is that the generic versions works for promoted cases, but the where
clause version does not. This is in support of extending `unorderedCopy`
to all POD types.