Fix a bug with taking the .type of a first class function
The definition of the class created to represent a first class function was
being inserted just before the call that it was created for. In the case
of the test added in this commit, this was inside of a type block. Type blocks
get removed after type resolution, but there were still references to the
first class function and the methods associated with it.

Instead, insert it just before the definition of the function that it
represents so that it will be visible from anywhere the function is visible.