Better determination of genericity of FnSymbols
Resolves #13721.

This PR makes two key changes.

#### Ensures computing the genericity of a FnSymbol before querying it

Switch from FnSymbol->hasFlag(FLAG_GENERIC) to FnSymbol->isGeneric().
This is so that we can assert that this has been computed.
If we are comfortable to drop the assertion, we can switch back.
Although I changed some conditionals from testing "is it generic" to
testing "have we calculated genericity". As a result, this is not
a 1:1 change.

For this switch, add two private fields in FnSymbol, one two indicate if it
is generic, the other to indicate whether the first has been calculated
intentionally. `isGeneric()` asserts that the second field is true.

Also add more calls to `FnSymbol::tagIfGeneric()`, which is the function
that computes genericity.

Remove `markGenericFunctions()`, that was invoked at start of resolution.
This is because it ran before formals' types were resolved and so could
miss a generic function. Instead, postpone computing genericity until
the function is examined for applicability to a call. That is, upon
resolveCall() --> filterCandidate() --> isApplicable(). See next.

#### Adds resolution of formal arguments' types before computing genericity

Before this change, a formal was sometimes mis-qualified as non-generic
because its type was not resolved yet. Causing a generic function to be
considered non-generic and leading to the resolution failure that was
worked around in #14166. See this for details:

    https://github.com/chapel-lang/chapel/issues/13721#issuecomment-535164115

Now tagIfGeneric() / hasGenericFormals() resolves the formal's type as needed.
I borrowed this from resolveSignature() / resolveFormals(). This is a step
towards running hasGenericFormals() and resolveFormals() simultaneously
for each formal, instead of running each for all of a function's formals
at once.

Ideally, we will also be detecting at the same time whether the function
is applicable to the CallExpr at hand and computing the bindings and the types
of already-instantiated formals. That way, if the function is found to be
not applicalbe due to one formal, we exit this process right away. Instead of
continuing to ponder whether the function is generic, which in this case
is irrelevant.

Notes:

* resolveFormals() does more than just resolve the `typeExpr`. However,
for this PR resolving the `typeExpr` is sufficient.

* hasGenericFormals() now completes as soon as it finds a generic formal,
instead of examining all formals always.

  - Otherwise, resolving a formal's type fails, for example, for the following
    declaration. There, we do not know `x(1).type` because we are not binding
    this FnSymbol to the call yet, so we do not know the type of `x`.
    Detecting applicability at the same time, discussed above, would remove
    this obstacle.

```chpl
      proc +(x: _tuple, y: x(1).type) ....
```

  - hasGenericFormals() used to detect whether there are generic formals
    without a default value. This required examining **all** formals.
    Given that this information has been unused, hasGenericFormals() does not
    compute it any longer.

  - When an `extern` function had multiple generic formals, the compiler would
    list them all in a single compilation. Now, only the first generic formal
    is reported.

* `tagIfGeneric()` is now called recursively on the same function sometimes,
so includes a guard that detects recursion and and aborts.

<details><summary>Here is what happens.</summary>

Consider the call `64*2` and these functions:
```chpl
proc *(a: real(?w), b: complex(w*2)) ...
proc *(a: imag(?w), b: complex(w*2)) ...
proc *(a: complex(?w), b: real(w/2)) ...
proc *(a: complex(?w), b: imag(w/2)) ...
```
which are cloned into:
```
proc *(a: real(32), b: complex(32*2)) ...
proc *(a: real(64), b: complex(64*2)) ...
proc *(a: imag(32), b: complex(32*2)) ...
......
```
Call them p1, p2, .... The compiler's steps are:
  - For each of the visible functions p1, p2, ..., is it applicable to this call?
  - Is p1 applicable?
  - Is p1 generic?
  - resolve the type `complex(32*2)`
  - resolve `32*2`
    - The visible functions are p1, p2, ... Check their applicability.
    - Is p1 applicable?
    - Is p1 generic?
    - We are already working on p1, abort the is-generic computation.
    - Condier p1 not applicable to this call since we aborted.
    - Is p2 applicable?
    - Is p2 generic?
    - resolve the type `complex(64*2)`
    - resolve the `64*2`
      - The visible functions are p1, p2, ... Check their applicability.
      - Now we skip p1, p2 and proceed with p3.
      - etc.

Thankfully this pattern has limited occurrences. Also, once a formal's type
has been resolved, the result is saved in the ArgSymbol, avoiding re-resolving
when this formal is encountered next time. I did not discern the effect of this
when compiling SSCA2. The compiler's `callStack` grows to 25 calls with the
above going on, however it is less than what we already get for SSCA2 on master.

</details>

* For those of us sensitive to the resolution order, note that the formal' type
may be resolved when checking the function's applicability to a call - during
tagIfGeneric(). Rather than when resolving the already-chosen best candidate.
I do not think that this is a change in behavior because we need to know
the formal's type in order to determine whether it applies. Indeed, the
compiler already calls resolveSignature() in isApplicableConcrete(). (Not sure
why it doesn't in isApplicableGeneric().)

#### Other

BTW compiler trivia: an FnSymbol with a type formal is considered non-generic
when it is used as the resolvedFunction() of a CallExpr. Even when the call
passes a generic type to that formal. This makes sense, first, because we
want CallExpr targets to be concrete, resolve them, etc. Second, at this point
all formals are instantiated as appropriate for this call and the FnSymbol
will not be instantiated any further.

While there:
* Switch from casts to CLASS_TYPE_UNMANAGED to casts to
  CLASS_TYPE_UNMANAGED_NONNIL to avoid unnecessary genericity.
* Remove a comment "it leaks memory" on a test that no longer does
  and an obsolete comment on resolveSignature().
* Improve the output format for vectors ofFnSymbols in view.cpp.

Remaining tasks:

* [ ] decide what to do with the the added `param` in compiler messages,
  see the changes in .good files - is this something to fix?
* [ ] add tests specifically for this change
* [ ] remove hack_resolve_types() (a stretch goal)
* [ ] retest linux64, gasnet multilocale tests