Move the RPM and module, for Shasta.
We've been using /opt as the base prefix for the Cray Chapel module, but
that's turning out to be problematic on Shasta systems.  Here, switch to
/opt/cray instead, on Shasta.  So the modules themselves will be in
/opt/cray/chapel and the modulefiles in /opt/cray/modulefiles/chapel.

A wrinkle is that we're also referring to /opt/cray/admin-pe/... for
some stuff that appears to be specific to XC and earlier systems.  I'm
checking on that, but it's working for now and doesn't have to be fixed
right away.