Update DistributedBag to use a rectangular array instead of associative
The array was meant to be a rectangular array with 'size' elements, but was
declared as var `A: [size] int`. When associative arrays allowed their domain
to be added to implicitly this worked.

A warning was added that this was deprecated, but the tests that use
DistributedBag use the '--quiet' flag, so they didn't hit the warning. Now
that it is a halt they do trip it.