Refactor multilocale libraries to compile client library with host compiler (#14248)
Refactor multilocale libraries to compile client library with host
compiler (#14248)

This PR adjusts multilocale library compilation so that the client
library is compiled by the CHPL_HOST_COMPILER instead of the
CHPL_TARGET_COMPILER.

It refactors the existing Makefiles into three separate files:

- Makefile.mli-common contains common variables and rules to
  compile the server
- Makefile.mli-shared contains rules to compile a shared ML
  client library with the host compiler
- Makefile.mli-static contains rules to compile a static ML
  client library with the host compiler

Use of recursive sub-makes was suggested by @bradcray as a way to
avoid modifying Makefile.include or other key Makefiles.

One downside of recursive make in this instance is that a lot of
variables originally injected by the compiler have to be explicitly
passed to the sub-make (or exported). Hopefully this downside is
mitigated by refactoring of common code.

I also modified two uses of uname, which is not generally portable,
to use CHPL_MAKE_PLATFORM instead.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=gasnet, HOST/TARG=gnu
- [x] interop/C/multilocale on darwin when CHPL_COMM=gasnet,
      HOST/TARG=clang
- [x] interop/python/multilocale on darwin (...)

For Crays, interop/C/multilocale/use_testing.ml-test.c, COMM=ugni:

- [x] HOST=gnu, TARG=PrgEnv-gnu
- [x] HOST=gnu, TARG=PrgEnv-intel
- [x] HOST=gnu, TARG=PrgEnv-cray

And as above, but for static libraries.

---

Thanks to @gbtitus for review!