Cleanup: CHPL_RPM --> rpmbuild_dir.
Remove mention of the environment variable CHPL_RPM from one script in
which it's neither set nor referenced.  In the other one that uses it,
rename it to rpmbuild_dir, because it's really a directory rather than
an RPM as the old name implied.

While here, also convert it to a shell variable because it's only used
within the one script.