Insert the promotion leader and follower iterators in a new location
Prior to this change, the promotion leader and follower iterators were getting
defined in chpl__Program, instead of defined next to the original function that
was getting promoted.  This meant that with the contents of ChapelBase no longer
moved to the program scope (due to the standardModuleSet "optimization"), we
could no longer find things like `chpl_here_alloc`.  Making them visible from
the scope where the symbol that is being promoted is defined should be
sufficient to avoid problems (but I will need to test this further to be
certain).