Fix unqualified uses of RangeChunk symbols in DefaultSparse
DefaultSparse had a `use RangeChunk only;`, which should have meant that the
symbols defined by RangeChunk could only be used in a qualified manner (e.g.
`RangeChunk.chunks`).  However, because the standardModuleSet was being used
before, we were able to get away with not doing that for these two references.
Now, they must be qualified, so do so.