Add a use of ChapelDebugPrint to these tests
These tests rely on chpl__testParStart, which is defined in ChapelDebugPrint and
only made available via private uses in the standard and internal modules.  In
order to access it here, we need an explicit use of that module now.  As a
result of the new use statement, the expected line numbers needed to be updated.