Don't place the promotion stuff at the fn's defPoint, place it slightly higher
Functions can be defined inside of functions, which results in the generated
function's scope being limited to inside a function (and thus not allowing it to
be found from outside of that function, which is problematic when it is used by
iterator records that are passed elsewhere).  Instead, place these functions
at the module scope of the symbol that caused their creation.