Fix bug in disambiguateByMatch
Basically, what was going on was that we were putting a candidate into the
ambiguous candidate set when it wasn't a sufficiently good match (but due to
other circumstances, we couldn't detect that it wasn't good enough).

We had 5 candidates, labelled 0-4.  Candidates 2 and 3 were the best matches,
but were equally as good as each other.  Candidates 0 and 4 were equally good as
each other, but less good than 2 and 3.  Candidate 1 was the worst match.
Basically, candidate 0 was getting marked as not good enough to be included
because it was encountered at the same time as candidate 2 (and later 3), but
when candidate 2 and 3 ran into each other, they caused an early end to the loop
before eliminating candidate 4 as an option.  When candidate 4 was considered,
it encountered candidate 0 (which it was equally as good as) and then stopped
looking for a match that was better than that.  But being as good as a notBest
match should mean that it is also a notBest candidate.  So propagate that
information when we know it (and hopefully we aren't able to get into another
situation like that).

So far as I can tell, we ran into this now because the order of candidates found
was changed.