Append library as last link arg in C compiles as well as LLVM (#14284)
Append library as last link arg in C compiles as well as LLVM (#14284)

This PR appends the user library as the last argument of the linker
invocation regardless of the chosen backend.

Certain linkers need some extra hand holding when resolving circular
references. We already used this workaround when compiling under
`--llvm`. Use it when compiling generated C as well.

Resolves #13457.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none and --llvm
- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @lydia-duncan for review!