Move definition of object class to ChapelBase from chpl__Program
The _new function for object relies on chpl_here_alloc, which is no longer
accessible from chpl__Program due to no longer moving everything included by
default into a single scope.  Since chpl__Program probably should not have a
use of ChapelStandard, make a definition of object in ChapelBase and use a
strategy to create it similar to that of locales, strings, and bytes.

Remove the majority of the code that created it from the compiler and replace
with other scaffolding.

I made the flags FLAG_OBJECT_CLASS and FLAG_GLOBAL_TYPE_SYMBOL available in
pragma form to support this change.  I will investigate removing
FLAG_GLOBAL_TYPE_SYMBOL from object (it might not be necessary, since the other
types I mentioned don't use it).