Adjust mpicc testing settings.
Use an OpenMPI module instead of an MPICH module for mpicc testing.

Always set CHPL_RT_OVERSUBSCRIBED for gasnet testing, not just with
qthreads tasking.  The tasking layer does matter in some ways, and we
make further adjustments based on that, but we'll be running multiple
processes per node no matter what.

When using mpicc and oversubscribed, throw the 'mpirun --map-by
node:oversubscribed' option, via MPIRUN_CMD.  This tells mpirun the
oversubscription is intentional and what to do about it.