Limit non-perf locModel=numa testing to just hellos and localeModels.
We don't need to be doing nearly so much locModel=numa functional
testing, given that we're also running performance testing on it.  So
here, reduce it to just the hellos and the localeModels suite.

As a beneficial side effect, making numa testing smaller will also make
it a lot simpler to think about moving it around, for load balancing or
other reasons.