Move the Error writeThis method to ChapelIO to avoid a circular dependency
ChapelError is required by SharedObject, which is initialized early in
ChapelStandard.  If ChapelError requires the IO module, it triggers the IO
module's initialization function before ChapelLocale's initialization function
has been run, meaning that its use of locale properties is not valid.  To
resolve this, move the method that requires the IO module to ChapelIO, which is
initialized after ChapelLocale.