Allow for forcing the number of NUMA domains, via the environment.
As an aid to testing, here allow a user to force the number of NUMA
domains the topology layer will say are present.  This is only effective
with locModel!=flat.

If tasks==qthreads this also sets CHPL_RT_NUM_THREADS_PER_LOCALE to the
same value, so that we have the same number of Qthreads shepherds as we
appear to have NUMA domains.  Otherwise, if Qthreads happened to choose
to run more shepherds we'd waste some of them, and if it happened to
choose to run fewer we'd get internal errors trying to queue to the
nonexistent ones.