Add a use of the IO module to fcf writeThis methods
With the IO module no longer included by default, the user might not have a use
of IO in their program, so trying to write a fcf might fail.  Add an explicit
use of IO to the FCF writeThis methods, so that they can always access the
symbols they need.