Throw -srequireThreadedMPI=false for all mpicc testing with OpenMPI.
If we're using MPI from OpenMPI, it appears we need to arrange to not
make concurrent MPI calls from multiple tasks/threads.  Do that here.

(We didn't/wouldn't need to do this with MPICH, but we switched away
from MPICH MPI in #14316 because we had trouble getting it to work with
our SLES12 testing systems.)