Remove fancy initializer that is not C++98 compliant (#14341)
Remove fancy initializer that is not C++98 compliant (#14341)

This PR removes a fancy "std::vector" initializer that is not C++98
compliant. I think that in this case I was trying to be too clever
for my own good.

I've also renamed one internal routine to be more to my liking.

---

Testing:

- [x] Build compiler with GCC v8.3, -std=c++98
- [x] ALL on linux64 when CHPL_COMM=none, CHPL_LLVM=none
- [x] interop/python on darwin when CHPL_COMM=none
- [x] interop/python on darwin when CHPL_COMM=gasnet

---

Thanks to @benharsh for review!