Improve performance for large transfers under `--cache-remote`
Bypass the cache when doing transfers that are bigger than 1/4 of the
cache. For large transfers we used to break the transfer into cached
sized sections. We'd add a section to the cache and immediately flush it
for the next section until all sections were done. This had performance
overheads and it limited the max RDMA size. e.g. instead of doing a
single 8GB transfer, we might have done a thousand 8M transfers, which
limited peak bandwidth.

Now we bypass the cache and directly initiate bulk comm, flushing any
pending operations that overlap with the bulk comm region first.

This resolves performance regressions for the arrayTransfer test, ISx,
and the blockToBlock exchange test.