Rename Shasta rpms as desired by delivery mechanisms.
For Shasta we need RPMs whose name includes a YYYYmmddHHMMSS timestamp
(in date(1) formatting terms) plus a 7-digit SHA for the branch from
which the RPM was built.  Here, add a '-R' option so the module build
driver can tell us that part of the name, and also arrange to create
it ourselves for Shasta RPMs if -R is not given.

While here, stop exporting some variables to the environment, because
they're just used locally and don't need to be inherited by other
scripts.

While here, also update some build documentation about naming and
installing RPMs.