Fix localSlice for cyclic and others
In response to private issue #538, this fixes the localSlice() routine
for Cyclic and propagates the same change to Block, which could be a
problem with non-naturally aligned strided Block arrays and domains.
I added a new distribution robustness test and in doing so found that
Replicated also didn't support localSlice so added a quick (but
seemingly correct) implementation for it as well.