Add a `locale.hostname` method
This adds a method named `hostname` to the `locale` type as a means of
letting users query the hostname that the locale represents / on which
it is running.  The existing `name` query typically returns the same
value, though in some cases we munge the hostname with additional
information or use different names (e.g., when using the local spawner
with certain GASNet configurations; for certain sub-locales).

The PR updates the locales primer to mention the method and adds a
test (for CHPL_COMM=none configurations only) which generates its
.good file dynamically using `hostname` from bash.