Tighten up prediff filter to restore test to previous state
In 1e23a753ec14a93c6d6bed43df80237c042a4a9e, Michael changed this
test's output to print the number of these() overloads being
considered during the resolution error.  This represents a
maintainability problem that wasn't intended as part of the test's
original scope as written.  For that reason, here I'm tightening
up the prediff filter to remove that line and restore the test
to its original behavior.