Add some tests of the error message for using the IO enums without a use of IO
A use of IO is now required to access iomode, iokind, iostringstyle, and
iostringformat.  We can interrupt the use of them in user code by utilizing
argumentless functions, but that prevents the enum constants from being usable.
So instead give a compilation error, since you won't be able to run it anyways