Convert .dim() from 1-based to 0-based indexing
I put this off in my original push to convert tuples to 0-based
thinking it should be a separate effort, but in retrospect, it
was inconsistent for me to convert other tuple-based queries on
domains and arrays (like .high / .low / etc.) without converting
this as well.  So I'm biting the bullet now.