Add CHPL_LIBFABRIC configuration setting.
For comm=ofi we've been referencing libfabric in an ad hoc manner, by
manually setting a provisional `LIBFABRIC_DIR` environment variable.
Here, add official support for a libfabric configuration setting, in the
form of a `CHPL_LIBFABRIC` environment variable which for now can be set
to either 'none' or 'system'.  The former is expected when comm!=ofi,
and the latter when comm=ofi.  When it is set to 'system', the scripts
expect to be able to find compile-time include-related options and
link-time library-related options via `pkg-config ... libfabric`.

In support of this, tighten up the libfabric API version specification
in the ofi comm layer proper.  For now we're insisting on having version
1.8 (the most recent).  We may relax this in the future, but doing so
would seem to imply making multiple `fi_getinfo()` calls looking for
supported versions, and I didn't feel like tackling that yet.

While here, fix a minor flaw in the error message printed when the
CHPL_MEM and CHPL_JEMALLOC settings conflict, which I spotted while
stealing script code in that area for CHPL_COMM and CHPL_LIBFABRIC.