Adjust failing future added in #14360
Adjust failing future added in #14360

This PR adjusts a future added in #14360 to better align it with my
original intent. A second future testing what happens _when_ the
override keyword is used with a subclass overload of `writeThis`
(where the parent `writeThis` is default generated) will be added
soon.

The future has also been marked as ".noexec", since it expects a
compiler error.

---

Changes to a future, trivial and not reviewed.