Clearly define lshift for signed int, add bounds checks for bitshift operators (#14391)
Clearly define lshift for signed int, add bounds checks for bitshift
operators (#14391)

This PR modifies the definitions for the left shift operator over
signed integers so that it is semantically equivalent to:

  (someSignedInt:uint(64) << shiftBy):int(64)

That is, the first argument is cast to its unsigned equivalent,
shifted, and then cast back.

It also adds bounds checks for bitshift operations that can be
disabled with use of the "--fast" flag:

  - Issue an error (or halt) if the amount to shift by is negative
  - Issue an error (or halt) if the amount to shift by is greater
    than or equal to the number of bits in the first operand

The first safety check is motivated by the Chapel standard. Both
safety checks prevent what is classified as "undefined behavior" by
the C standard.

Resolves #14305.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none and CHPL_LLVM=system
- [x] ALL on linux64 when CHPL_COMM=gasnet and CHPL_LLVM=system
- [x] expressions/bitshift on linux32 when CHPL_COMM=none

---

Thanks to @e-kayrakli for review!
Thanks to @cassella for catching a silly performance bottleneck.