Add a temporary module for IO deprecation warnings
Adds the module LastResortIO.  This module contains functions with the same name
as symbols defined by the IO module, which generate a deprecation warning
indicating that the symbols are no longer available by default and then performs
the original action.  In some cases, it is not possible to enable the original
action after the message, and so a deprecation error is used instead.

The symbols it covers are:
    - openmem
    - openreader
    - opentmp
    - openfp
    - openfd
    - open
    - iohints
    - defaultIOStyle
    - IOHINT_*
    - stringStyleWithVariableLength
    - stringStyleWithLength
    - stringStyleExactLen
    - stringStyleNullTerminated
    - iolittle
    - iobig
    - ionative
    - iodynamic
    - iokind
    - iostringformat
    - iostringstyle
    - iomode
and briefly file & file.home

This file is a work in progress - I have TODOs for deprecation warnings I have
not tried to implement yet, and have commented out the deprecation warnings for
symbols I have not yet tested, due to concerns about generating unhelpful
messages.  I intend to continue working on this but don't think it is reasonable
to hold up the current PR until they are finished, as we hope it will recover
some of the increased compilation time experienced as a result of removing the
standardModuleSet "optimization"