Update these tests and MasonDoc for the new deprecation warnings
The mason files were getting away with the fact that modules they use had a
chain that led to a public use of IO.  However, they did not have an explicit
use themselves.  Because this use happened to be at a similar distance to the
use of LastResortIO, the deprecation function and the enum were in conflict,
instead of having one or the other take precedence.

The .bad files for the shootouts were updated to reflect the new state.  All
were failing to compile due to not finding a symbol defined in the IO module.
Now, some of them still fail to compile, but also tell you to explicitly use the
IO module.  Others of them no longer yell about an IO symbol, but still yell
early about other symbols (so never reach the deprecation warning).  And one of
them now warns about needing a use of IO and then compiles and runs.