Deal with odd performance regression
When changing `use` statements to `private use` by default in user
code, the rvfSlices/reduceSlice.chpl test took a performance hit
because some canResolve() calls in chpl__rvfMe() in
ArrayViewSlice.chpl were returning false where they've typically
returned true.  Making the `use` in reduceSlice.chpl `public`
preserves the old values.  This would seem to relate to whether or not
methods on a type are available if the module that defined that type
is not currently public (?).

For now, I've changed the `use` to `public` explicitly in the code and
filed a clone of the test to capture the public vs. private
difference.  I've set up the new test to skipif for anything other
that a fairly default CHPL_COMM!=none, default distribution scenario
to avoid having to make all the various .good files.  I'll also check
into creating a simpler, standalone version of the test.