Fix a simple bug that had crept into ChapelIteratorSupport
This fixes a bug that crept into ChapelIteratorSupport in which we
were returning the 'none' value in a function that is supposed to
return types.  Here, rather than changing the type into 'nothing', I
changed it back to the original 'void' since (a) it matches the
original function name's statement of what it's doing and (b) the type
is used as a sentinel value, so it doesn't particularly matter that
it's a type that has no values.