Drop warning for open function
This function has a required argument that is of a type which will cause the
user's program to fail to compile.  Thus, it is not necessary to add a warning
for it, as the compilation error will have already alerted the user of their
need for a use of the IO module.

As a result, remove the check of the compilation error from this test.