Make the flag that says whether to init ddatas a param
At present, we always know at compile-time whether or not a ddata
wants to be initialized or not, so change the flag that indicates this
into a param such that the compiler will fold out the initialization
code.  Failure to do so causes a problem when trying to create a ddata
of types that don't have an initial value, like non-nilable classes.
Add a test to lock this behavior in.

In the future, if/when we have a case that needs this not to be a param,
we can presumably add a non-param overload and have the list continue
to use the param overload as it is today (?).