Add uses of the IO module to Curl, URL, and some tests that use them
URL depends on IO symbols as part of the exposed interface for
openUrlReader, so needs a public use of IO.

Curl depends on IO symbols as part of the exposed interface for
getCurlHandle, so needs a public use of IO.  It also has a submodule
that is undocumented and uses IO symbols, so I made the use there
private.

The tests make use of IO symbols directly in ways that aren't related
to Curl and URL itself, so they should have an explicit use.