Make CommDiagnostics no longer included by default
A lot of tests in our test system have an explicit use of CommDiagnostics and
the documentation seems to think it must be used.  However, due to its public
uses in start and stopInitCommDiags, and the public use of those modules in
ChapelStandard, it was actually getting included in all programs by default.

This change makes it so that those uses are private, enforcing the behavior that
we were already intending to have.