Make the use of MemConsistency in ChapelLocks private
This module utilizes MemConsistency's memoryOrders, but it does so as part of
their implementation.  As a result, the use of MemConsistency does not need to
be public.

Note: the use of Atomics in this module maybe should be public rather than
private.  chpl_LocalSpinlock's field does use a symbol from it for its type,
but it's possible that field could become private once private fields are
implemented.