Make some uses in ChapelDistribution intentionally public
ChapelLocks is used so that the _domsLock field's type can be used in calls to
its initializer.

LinkedLists are used so that the _doms field's type can be used in calls to its
initializer.

These field names start with an underscore, so are potential private field
candidates when that support exists.  If the fields become private, these use
statements could also potentially become private.