Add a private use of ChapelBase to all the locale models
For flat, I tried to narrow that down somewhat to prevent future effort (I think
we will eventually narrow what about ChapelBase is included by default, so any
uses I can add instead of just punting the problem until then will be good).

It looks like it needed a use of ChapelRange to build a bounded range and a use
of ChapelIteratorSupport to call _getIterator.  With that, it still needed a use
of ChapelBase to assign between bools.