Allow throws in task-private variables' initializing expressions
Resolves #13203. Resolves #14386.

This extends the following steps in compiler's lowering of error handling
to apply to ShadowVarSymbols' init/deinit blocks:

* detection of whether those blocks may throw,
* propagation of the `error` argument.

While there, a minor fix to a chpldoc and remove a mistakenly-added file.

Next steps:

* Fix the line number for errors thrown within the init expression
for a task-private variable in the outermost task. This is the task
that encounters the forall statement and that executes the parallel iterator
before it enters any of its own task constructs. Currently the TaskErrors
instance whose line number is missing is created within `chpl_forall_error()`
with `return new unmanaged TaskErrors(err)`. Cf. the TaskErrors instance
created for errors thrown in child tasks - whose line number is meaningful -
comes from `_waitEndCount()` - `throw new owned TaskErrors(e.errors)`.

* Execute task-private variables' **deinitializers** appropriately
when one of their intializing expressions throws. Currently the generated code
would execute all the deinitializers in such a case, leading to deinit of
uninitialized variable(s). To prevent this, currently the compiler emits
an error when any of the deinitializers involved are potentially non-POD.
This is #9749.